feat: add GET endpoint for getting students data
Add default parameter to getFileJson(), because we want this function to not truncate json file in some situations.

+++ api.js
@@ -1,7 +1,8 @@
 const fs = require('fs')
 
 module.exports = {
-  storeStudentData
+  storeStudentData,
+  getStudentData
 }
 
 const fsPromises = fs.promises
@@ -13,7 +14,7 @@ async function storeStudentData (req, res, next) {
   const path = `./data/${studentId}.json`
 
   try {
-    let json = await getFileJson(path)
+    let json = await getFileJson(path, true)
     checkNestedProperties(json, properties)
     addProperty(json, properties, body)
     await storeJson(path, json)
@@ -24,8 +25,24 @@ async function storeStudentData (req, res, next) {
   }
 }
 
-async function getFileJson (path) {
-  if (!fs.existsSync(path)) {
+async function getStudentData (req, res, next) {
+  const { studentId } = req.params
+  const properties = req.params[0].split('/')
+  const path = `./data/${studentId}.json`
+
+  try {
+    let json = await getFileJson(path)
+    let property = getNestedProperty(json, properties)
+
+    res.json({ success: true, body: property })
+  } catch (e) {
+    res.status(404)
+    res.json({ success: false, error: e })
+  }
+}
+
+async function getFileJson (path, createIfDoesNotExist = false) {
+  if (createIfDoesNotExist && !fs.existsSync(path)) {
     await fsPromises.writeFile(path, '{}')
   }
 
@@ -40,6 +57,13 @@ function checkNestedProperties (json, properties) {
     }
     json = json[property]
   }
+}
+
+function getNestedProperty (json, properties) {
+  for (const property of properties) {
+    json = json[property]
+  }
+  return json
 }
 
 function addProperty (json, properties, lastObject) {
